Here is a second round of cleanup and fixes for the realtime testcases.
1) Make sched_jitter use the create_fifo_thread() library function instead of an open coded solution,
2) Prio-wake calls rt_init() twice, remove the second call,
3) Make sbrk_mutex less verbose by default. One can still use the -v option to get the whole output,
4) It's better to calculate the histogram before saving it. This was introduced in an earlier commit of mine fixing the quantile calculation,
5) Fix runtime displaying of the min and max latencies (when used with -v3). While at it, remove an uneeded avg variable,
6) Various tests still have a hardcoded value for the quantile nines. Use a value automatically calculated from the number of iterations,
7) The log10() call used for automatic quantile nines calculation returns a double result. Cast it to an int. The exp10() call used in stats_quantiles_calc() for checking purposes returns a double result which is compared against a long. Cast it to a long. This allows the following comparison: data->size < (long)exp10(quantiles->nines) to really be false when quantiles->nines has been calculated as log10(data->size).
More generally, it seems that (at least with gcc 4.1.1):
        long i = 10000;
        double f = exp10(log10(i))

        yields (i < f) being true due to rounding,
8) Add latency tracing capability to pthread_kill_latency as is already done on a few other latency tests (gtod_latency, sched_latency, ...),
9) The '::' optional argument specifier for getopt used by the '-v' option is a GNU extension, is not portable and does not work. For example it's not even described in the Debian getopt(3) manpage. Make the '-v' option require a non optional argument,
10)The print buffer is only ever flushed when it is full. Add flushing when the test terminates vi atexit(),
11)The 'period missed' check of the thread first loop should not depend on the thread starting time. This is especially visible on 'slow' platforms where one cannot run the test if thread creation takes a long time. Fix it by removing this dependency. All delays are now calculated relative to when the thread starts,

Cc: Darren Hart <dvhltc@us.ibm.com>
Cc: Tim Chavez <tinytim@us.ibm.com>
Cc: Matthieu CASTET <matthieu.castet@parrot.com>
Acked-by: Chirag <chirag@linux.vnet.ibm.com>

+++ testcases/realtime/func/periodic_cpu_load/periodic_cpu_load.c
@@ -196,7 +196,7 @@ int main(int argc, char *argv[])
 
 	for (i=0; i<(THREADS_PER_GROUP * NUM_GROUPS); i++) {
 		stats_container_init(&dat[i], ITERATIONS);
-		stats_quantiles_init(&quantiles[i], log10(ITERATIONS));
+		stats_quantiles_init(&quantiles[i], (int)log10(ITERATIONS));
 	}
 
 	struct periodic_arg parg_a = {PERIOD_A, ITERATIONS, calc, (void *)CALC_LOOPS_A };
